+++ kernel/CommandLine.cpp
@@ -17,7 +17,6 @@
     Copyright (c) Byteduck 2016-2020. All rights reserved.
 */
 
-#include <kernel/memory/PageDirectory.h>
 #include <kernel/kstd/KLog.h>
 #include "CommandLine.h"
 #include "kernel/memory/MemoryManager.h"
