+++ kernel/api/stdarg.h
@@ -7,19 +7,12 @@
 
 __DECL_BEGIN
 
-typedef char* va_list;
+typedef __builtin_va_list va_list;
 
-#define __va_argsiz(t)	\
-	(((sizeof(t) + sizeof(int) - 1) / sizeof(int)) * sizeof(int))
+#define va_start(ap, pN) __builtin_va_start(ap, pN)
 
+#define va_end(ap) __builtin_va_end(ap)
 
-#define va_start(ap, pN)	\
-	((ap) = ((va_list) __builtin_next_arg(pN)))
-
-#define va_end(ap)	((void)0)
-
-#define va_arg(ap, t)					\
-	 (((ap) = (ap) + __va_argsiz(t)),		\
-	  *((t*) (void*) ((ap) - __va_argsiz(t))))
+#define va_arg(ap, t) __builtin_va_arg(ap, t)
 
 __DECL_END
