+++ kernel/constructors.h
@@ -1,8 +1,6 @@
 /* SPDX-License-Identifier: GPL-3.0-or-later */
 /* Copyright © 2016-2024 Byteduck */
 
-#include "rpi/MiniUART.h"
+#pragma once
 
-void serial_putch(char c) {
-//	RPi::MiniUART::tx(c);
-}
+void call_global_constructors();
