+++ kernel/memory/Memory.h
@@ -34,7 +34,7 @@ extern "C" long _PAGETABLES_END;
 #define KERNEL_TEXT_SIZE (KERNEL_TEXT_END - KERNEL_TEXT)
 #define KERNEL_DATA_SIZE (KERNEL_DATA_END - KERNEL_DATA)
 #define KERNEL_END_VIRTADDR (HIGHER_HALF + KERNEL_SIZE_PAGES * PAGE_SIZE)
-#define KERNEL_VIRTUAL_HEAP_BEGIN 0xE0000000
+#define KERNEL_VIRTUAL_HEAP_BEGIN (HIGHER_HALF + 0x20000000)
 #define MAX_QUICKMAP_PAGES 8
 #define KERNEL_QUICKMAP_PAGES (KERNEL_VIRTUAL_HEAP_BEGIN - (PAGE_SIZE * MAX_QUICKMAP_PAGES))
